% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_units_compatible.R
\name{all_units_compatible}
\alias{all_units_compatible}
\title{Test if a vector of units are all compatible.}
\usage{
all_units_compatible(x)
}
\arguments{
\item{x}{[character]}
}
\value{
\code{all_units_compatible}: [Logical(1)]
}
\description{
Test if a vector of units are all compatible.
}
\examples{
 x = c('in', 'inches', 'inch')
 all_units_compatible(x)
 x = c('ft', 'inches', 'm')
 all_units_compatible(x)
 x = c('ft', 'inches', 'kg')
 all_units_compatible(x)
}
