% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_units_identical.R
\name{all_units_identical}
\alias{all_units_identical}
\title{Test if a vector of units are all identical units.}
\usage{
all_units_identical(x)
}
\arguments{
\item{x}{[character]}
}
\value{
\code{all_units_identical}: [Logical(1)]
}
\description{
Test if a vector of units are all identical units.
}
\examples{
 x = c('in', 'inches', 'inch')
 all_units_identical(x)
 x = c('ft', 'inches', 'm')
 all_units_identical(x)
 x = c('ft', 'inches', 'kg')
 all_units_identical(x)
}
