% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_unit_name.R, R/check_unit_name.R
\name{assert_unit_name}
\alias{assert_unit_name}
\alias{check_unit_name}
\title{Assert if a string is a valid unit name.}
\usage{
assert_unit_name(
  x,
  null.ok = FALSE,
  .var.name = checkmate::vname(x),
  add = NULL
)

check_unit_name(x, null.ok = FALSE)
}
\arguments{
\item{x}{an object to check}

\item{null.ok}{[logical(1)] If set to TRUE, x may also be NULL.}
}
\value{
\code{assert_unit_name}: If valid, returns x invisibly. If invalid, throws an error

\code{check_unit_name}: TRUE if valid. If invalid, returns a string with an error message.
}
\description{
Assert if a string is a valid unit name.

Check if a string is a valid unit name.
}
\examples{
 assert <- assert_unit_name('m')
 assert
 assert_unit_name('mk')
 check <- check_unit_name('m')
 check
 check_unit_name('mk')
}
