% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unitable.R
\name{check_unitable}
\alias{check_unitable}
\title{Check if an object is unitable.}
\usage{
check_unitable(x)
}
\value{
\code{check_unitable}: TRUE if valid. If invalid, returns a string with an error message.
}
\description{
Check if an object is unitable.
}
\examples{
 check_unitable(1)
 check_unitable(1)
 check_unitable(1)
 check_unitable('a')
 check_unitable(factor('a'))
 check_unitable(data.table())
 check_unitable(Sys.time())
 check_unitable(Sys.Date())
}
