% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_unit_class_formula.R
\name{edit_unit_class_formula}
\alias{edit_unit_class_formula}
\title{Edit a base unit to dataable}
\usage{
edit_unit_class_formula(unit_name, formula)
}
\arguments{
\item{unit_name}{the unit name}

\item{formula}{the formula in langauge format}
}
\value{
Invisibly returns the base_units DB
}
\description{
Edit a base unit to dataable
}
