% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compatible_unit_vars.R,
%   R/set_compatible_unit_vars.R
\name{get_compatible_unit_vars}
\alias{get_compatible_unit_vars}
\alias{set_compatible_unit_vars}
\title{Get or set all numeric variables base on units.}
\usage{
get_compatible_unit_vars(DT, return = "list")

set_compatible_unit_vars(DT)
}
\arguments{
\item{DT}{[data_table]}

\item{return}{[choice('list|DT')] Should the function return a named list or data.table?}
}
\value{
\code{get_compatible_unit_vars}: A list of numeric variable names grouped by unit.

\code{set_compatible_unit_vars}: invisible([list]) A list of numeric variable names grouped by unit.
}
\description{
Groups all numeric variables in a data_table base on compatible unit.
}
\examples{
 non_equal = sapply(DT, get_units, require.numeric = FALSE)
 set_compatible_unit_vars(DT)
 equal = sapply(DT, get_units, require.numeric = FALSE)
 rbind(non_equal, equal)
}
