% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unit_aliases.R
\name{get_unit_aliases}
\alias{get_unit_aliases}
\title{Get all acceptable aliases of a unit name.}
\usage{
get_unit_aliases(name)
}
\arguments{
\item{name}{[character]  Must have an exact length of 1.}
}
\value{
[Character] vector.
}
\description{
Get all acceptable aliases of a unit name.
}
\examples{
 get_unit_aliases("lb")
}
