% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unit_names.R
\name{get_unit_names}
\alias{get_unit_names}
\title{Get a vector of all unit names in the database.}
\usage{
get_unit_names(type = c("base", "derived", "accepted", "common", "prefixes"))
}
\arguments{
\item{type}{[subset]  Possible values: c("base", "derived", "accepted", "common").  Defaults to c("base", "derived", "accepted", "common")}
}
\value{
[Character] vector.
}
\description{
Get a vector of all unit names in the database.
}
\examples{
 get_unit_names()
}
