% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_units.R
\name{get_units}
\alias{get_units}
\title{Extract units from a numeric vector.}
\usage{
get_units(x, require.numeric = TRUE)
}
\arguments{
\item{x}{\code{[numeric]}}

\item{require.numeric}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}
}
\value{
\code{get_units}: \code{[character]} the unit type
}
\description{
Extract units from a numeric vector.
}
\examples{
 x <- set_units(1, 'km*W/s^2/m^2')
 get_units(x)
 x <- set_units(1, 'mi/gallon')
 get_units(x)
 x <- set_units(1)
 get_units(x)
 x <- set_units(1, '1/s')
 get_units(x)
 x = 'A'
 get_units(x, require.numeric = FALSE)
}
