% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_units.R
\name{has_units}
\alias{has_units}
\title{Test if a numeric vector has defined units.}
\usage{
has_units(x)
}
\arguments{
\item{x}{[numeric]}
}
\value{
Logical
}
\description{
Test if a numeric vector has defined units.
}
