% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_convertable_unit_pair.R
\name{is_convertable_unit_pair}
\alias{is_convertable_unit_pair}
\title{Check if a unit pair is convertible}
\usage{
is_convertable_unit_pair(x, y)
}
\arguments{
\item{x}{A character string which is parseable into a udunits compatible unit.}

\item{y}{Another character string which is also parseable into a udunits compatible unit.}
}
\value{
Logical. True if unit pair is convertable
}
\description{
Check if a unit pair is convertible
}
