% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unit.R
\name{is_unit}
\alias{is_unit}
\title{Test if a numeric vector is a specific unit class.}
\usage{
is_unit(x, unit_class)
}
\arguments{
\item{x}{[numeric]}

\item{unit_class}{[choice]  Possible values: \code{c( units_classes()$unit, 'count', 'unitless')}.}
}
\value{
\code{is_unit}: [Logical(1)]
}
\description{
Test if a numeric vector is a specific unit class.
}
\examples{
 x = set_units(1, 'm/s')
 y = set_units(1, 'm^2')
 z = set_units(1)
 is_unit(x, 'speed')
 is_unit(y, 'area')
 is_unit(z, 'area')
 is_unit(z, 'unitless')
 is_unit(1, 'unitless')
}
