% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unit_demoninator.R
\name{is_unit_demoninator}
\alias{is_unit_demoninator}
\title{Test if a numeric vector's demoninator is of a certain unit class.}
\usage{
is_unit_demoninator(x, unit_class)
}
\arguments{
\item{x}{[numeric]}

\item{unit_class}{[choice]}
}
\value{
Logical
}
\description{
Test if a numeric vector's demoninator is of a certain unit class.
}
