% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unit_numerator.R
\name{is_unit_numerator}
\alias{is_unit_numerator}
\title{Test if a numeric vector numerator is of a certain unit class.}
\usage{
is_unit_numerator(x, unit_class)
}
\arguments{
\item{x}{[numeric]}

\item{unit_class}{[choice]}
}
\value{
Logical
}
\description{
Test if a numeric vector numerator is of a certain unit class.
}
