% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unit_user_defined.R
\name{is_unit_user_defined}
\alias{is_unit_user_defined}
\title{Determine if a unit is user defined.}
\usage{
is_unit_user_defined(unit_name)
}
\arguments{
\item{unit_name}{[character]  Must have an exact length of 1.}
}
\value{
[Logical]
}
\description{
Determine if a unit is user defined.
}
\examples{
 is_unit_user_defined("kg")
}
