% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unitable.R
\name{is_unitable}
\alias{is_unitable}
\title{Determine if an object can have measurement units.}
\usage{
is_unitable(x)
}
\arguments{
\item{x}{[R object]}
}
\value{
\code{is_unitable}: [Logical(1)]
}
\description{
Determine if an object can have measurement units.
}
\examples{
is_unitable(list('a', 'b', 'c'))
is_unitable(data.table(3, 2, 1))
is_unitable(Sys.time())
is_unitable(Sys.Date())
is_unitable(1L:10L)
is_unitable(NA_real_)
is_unitable(NA_integer_)
is_unitable(NA_complex_)
is_unitable(NA_character_)
is_unitable(1.0)
}
