% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_unit_name.R
\name{is_valid_unit_name}
\alias{is_valid_unit_name}
\title{Test whether a string is a valid unit name.}
\usage{
is_valid_unit_name(x)
}
\arguments{
\item{x}{[character(1)]}
}
\value{
[Logical]
}
\description{
Test whether a string is a valid unit name.
}
\examples{
 is_valid_unit_name('m')
 is_valid_unit_name('J')
 is_valid_unit_name('fish')
}
