% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_units.R
\name{map_units}
\alias{map_units}
\title{Set or drop units across a list, data.table, or, data.frame}
\usage{
map_units(x, units = NULL, drop = FALSE, force_list = FALSE)
}
\arguments{
\item{x}{[data.table,data.frame,list]  Each column or list element must inherit from at least one of the provided types: [numeric]}

\item{units}{[character]  Must have an exact length of or equal to one of the following: [1,length(x)].  NULL is ok.  Defaults to NULL}

\item{drop}{[logical] Drop existing units? Defaults to FALSE}

\item{force_list}{[logical]  Return as a list? Defaults to FALSE}
}
\value{
[list,data.table,data.frame] depending on the input class. Returns as a list if specified by 'force_list'.
}
\description{
Set or drop units across a list, data.table, or, data.frame
}
\examples{
 LST = list(x = rnorm(10), y = rnorm(20), z = rnorm(5))
 map_units(LST, units = c('m/s'))
 DT = data.table(x = rnorm(50), y = rnorm(50), z = rnorm(50))
 DT[, `:=`(c('x', 'y', 'z'), map_units(.SD, units = c('m/s')))]
 DT
DF = as.data.frame(DT)
map_units(DF, units = c('m'), drop = TRUE)
DT = data.table(x = rnorm(50), y = rnorm(50), z = rnorm(50))
map_units(DT)
}
