% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_common_unit.R
\name{new_common_unit}
\alias{new_common_unit}
\title{Create a new common unit.}
\usage{
new_common_unit(name, formula, aliases, symbol = NULL, definition = NULL)
}
\arguments{
\item{name}{[character]  Must have an exact length of 1.}

\item{formula}{[character]  Must have an exact length of 1.}

\item{aliases}{[character]  NULL is ok.}

\item{symbol}{[character]  NULL is ok.  Defaults to NULL}

\item{definition}{[character]  Must have an exact length of 1.  NULL is ok.  Defaults to NULL}
}
\value{
[Numeric] class unit
}
\description{
Common units have formulas derived from base units.
}
\examples{
 new_base_unit(name = "USD", aliases = c("dollar", "us_dollar"),
 symbol = "$", definition = "Official currency of the United States")
 new_common_unit(name = "EUR", formula = "1.18*dollar", aliases = "euro",
 symbol = "\200", definition = "Official currency of the European Union")
 remove_user_defined_unit("USD", remove_dependencies = TRUE)
}
