% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_unit_catagory.R
\name{new_unit_class}
\alias{new_unit_class}
\title{Add a new base unit class to database}
\usage{
new_unit_class(unit_name, formula)
}
\arguments{
\item{unit_name}{the unit name}

\item{formula}{the formula in langauge format}
}
\value{
Invisibly returns the unit_class DB
}
\description{
Add a new base unit class to database
}
