% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unit_class.R
\name{remove_unit_class}
\alias{remove_unit_class}
\title{Remove a unit class from unit_class database}
\usage{
remove_unit_class(unit_name)
}
\arguments{
\item{unit_name}{the unit name}
}
\value{
Invisibly returns the unit_class DB
}
\description{
Remove a unit class from unit_class database
}
