% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_user_defined_unit.R
\name{remove_user_defined_unit}
\alias{remove_user_defined_unit}
\title{Remove a user defined unit.}
\usage{
remove_user_defined_unit(unit_name, remove_dependencies = FALSE)
}
\arguments{
\item{unit_name}{[character]  Must have an exact length of 1.}

\item{remove_dependencies}{[logical]  Defaults to FALSE}
}
\value{
[NULL]
}
\description{
Remove a user defined unit.
}
\examples{
 new_base_unit(name = "USD", aliases = c("dollar", "us_dollar"),
 symbol = "$", definition = "Official currency of the United States")
 new_common_unit(name = "EUR", formula = "1.18*dollar", aliases = "euro",
 symbol = "\200", definition = "Official currency of the European Union")
 remove_user_defined_unit("USD", remove_dependencies = TRUE)
}
