% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_unit_class.R
\name{rename_unit_class}
\alias{rename_unit_class}
\title{Rename a unit class}
\usage{
rename_unit_class(old_name, new_name)
}
\arguments{
\item{old_name}{the old unit name}

\item{new_name}{the new unit name}
}
\value{
Invisibly returns the unit_class DB
}
\description{
Rename a unit class
}
