% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_equal_units.R
\name{set_equal_units}
\alias{set_equal_units}
\title{Set R vectors with convertible units to the same unit.}
\usage{
set_equal_units(x, unit = NULL)
}
\arguments{
\item{x}{[data.table,list,data.frame]  Each column or list element must inherit from at least one of the provided types: [units]}

\item{unit}{[unit_name]  NULL is ok.  Defaults to NULL}
}
\value{
\code{set_equal_units}: [data.table,data.table,or list] depending on input class
}
\description{
If argument unit is null, the unit of the first vector in the list will be used.
}
\examples{
 DT <- data.table(x = 1:5, y = 2:6, z = rnorm(5))
 DT = map_units(DT, units = c('m/s', 'km/hr', 'mi/hr'))
 set_equal_units(DT)
 set_equal_units(DT[, .(z, y, x)])
 set_equal_units(DT[, .(z, y, x)], unit = 'ft/min')
}
