% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_class.R
\name{unit_class}
\alias{unit_class}
\title{Determine the unit class of a numeric vector.}
\usage{
unit_class(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the unit class.  If the vector has units and no class is determined, a formula string will be returned.
}
\description{
Determine the unit class of a numeric vector.
}
\examples{
 x <- set_units(1, "m/s")
 unit_class(x)
 x <- set_units(1, "m/s^4")
 unit_class(x)
 unit_class(1)
}
