% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_denominator.R
\name{unit_denominator}
\alias{unit_denominator}
\title{Get the unit denominator of an numeric vector.}
\usage{
unit_denominator(x)
}
\arguments{
\item{x}{[numeric]}
}
\value{
[Character] The denominator formula
}
\description{
Get the unit denominator of an numeric vector.
}
