% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_has_denominator.R
\name{unit_has_denominator}
\alias{unit_has_denominator}
\title{Test if a numeric vector has a denominator.}
\usage{
unit_has_denominator(x)
}
\arguments{
\item{x}{[numeric]}
}
\value{
Logical
}
\description{
Test if a numeric vector has a denominator.
}
