% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_has_numerator.R
\name{unit_has_numerator}
\alias{unit_has_numerator}
\title{Test if a numeric vector has a numerator.}
\usage{
unit_has_numerator(x)
}
\arguments{
\item{x}{[numeric]}
}
\value{
Logical
}
\description{
Test if a numeric vector has a numerator.
}
