% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_numerator.R
\name{unit_numerator}
\alias{unit_numerator}
\title{Get the unit numerator of an numeric vector.}
\usage{
unit_numerator(x)
}
\arguments{
\item{x}{[numeric]}
}
\value{
[Character] The numerator formula
}
\description{
Get the unit numerator of an numeric vector.
}
