% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_units_compatible.R
\name{which_units_compatible}
\alias{which_units_compatible}
\title{Finds all compatible units within a vector of units.}
\usage{
which_units_compatible(x, table)
}
\arguments{
\item{x}{[character]  Must have an exact length of 1.}

\item{table}{[character]}
}
\value{
\code{which_units_compatible}: [Logical] All units compatible with x.
}
\description{
Finds all compatible units within a vector of units.
}
\examples{
 x = 'm'
 table = c('km', 'J', 'W', 'N', 'r', 's', 'day')
 which_units_compatible(x, table)
}
